refactor: Move components and validations to folders named groups
