+++ src/app/[locale]/(default)/about/group/new/page.tsx
@@ -1,7 +1,7 @@
 import { ArrowLeftIcon } from 'lucide-react';
 import { type Messages, NextIntlClientProvider } from 'next-intl';
 import { getMessages, getTranslations } from 'next-intl/server';
-import { EditGroupForm } from '@/components/about/EditGroupForm';
+import { EditGroupForm } from '@/components/groups/EditGroupForm';
 import { Link } from '@/components/ui/Link';
 import { api } from '@/lib/api/server';
