+++ src/validations/groups/editGroupSchema.ts
@@ -1,6 +1,6 @@
 import { z } from 'zod';
 import type { Translations } from '@/lib/locale';
-import { groupSchema } from '@/validations/about/groupSchema';
+import { groupSchema } from '@/validations/groups/groupSchema';
 
 function editGroupSchema(t: Translations) {
   return groupSchema(t).extend({
